+++ sdk/Sdk.Generators/FunctionExecutor/FunctionExecutorGenerator.Emitter.cs
@@ -33,6 +33,7 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
                          namespace {{FunctionsUtil.GetNamespaceForGeneratedCode(context)}}
                          {
                              [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
+                             [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
                              internal class DirectFunctionExecutor : IFunctionExecutor
                              {
                                  private readonly IFunctionActivator _functionActivator;{{(defaultExecutorNeeded ? $"{Environment.NewLine}        private Lazy<IFunctionExecutor> _defaultExecutor;" : string.Empty)}}