+++ test/Sdk.Generator.Tests/FunctionMetadataProviderGeneratorTests/HttpTriggerTests.cs
@@ -77,6 +77,7 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                     /// Custom <see cref="IFunctionMetadataProvider"/> implementation that returns function metadata definitions for the current worker."/>
                     /// </summary>
                     [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
+                    [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
                         /// <inheritdoc/>
@@ -171,6 +172,7 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                     /// Custom <see cref="IFunctionMetadataProvider"/> implementation that returns function metadata definitions for the current worker."/>
                     /// </summary>
                     [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
+                    [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
                         /// <inheritdoc/>
@@ -269,6 +271,7 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                     /// Custom <see cref="IFunctionMetadataProvider"/> implementation that returns function metadata definitions for the current worker."/>
                     /// </summary>
                     [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
+                    [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
                         /// <inheritdoc/>