+++ test/Sdk.Generator.Tests/FunctionMetadataProviderGeneratorTests/IntegratedTriggersAndBindingsTests.cs
@@ -1,11 +1,9 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-using System.Collections.Generic;
 using System.Reflection;
 using System.Threading.Tasks;
 using Microsoft.Azure.Functions.Worker.Sdk.Generators;
-using Microsoft.CodeAnalysis.Testing;
 using Xunit;
 
 namespace Microsoft.Azure.Functions.SdkGeneratorTests
@@ -112,6 +110,7 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                     /// Custom <see cref="IFunctionMetadataProvider"/> implementation that returns function metadata definitions for the current worker."/>
                     /// </summary>
                     [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
+                    [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
                         /// <inheritdoc/>
@@ -247,6 +246,7 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                     /// Custom <see cref="IFunctionMetadataProvider"/> implementation that returns function metadata definitions for the current worker."/>
                     /// </summary>
                     [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
+                    [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
                         /// <inheritdoc/>
@@ -357,6 +357,7 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                     /// Custom <see cref="IFunctionMetadataProvider"/> implementation that returns function metadata definitions for the current worker."/>
                     /// </summary>
                     [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
+                    [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
                         /// <inheritdoc/>
@@ -452,6 +453,7 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                     /// Custom <see cref="IFunctionMetadataProvider"/> implementation that returns function metadata definitions for the current worker."/>
                     /// </summary>
                     [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
+                    [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
                         /// <inheritdoc/>
@@ -546,6 +548,7 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                     /// Custom <see cref="IFunctionMetadataProvider"/> implementation that returns function metadata definitions for the current worker."/>
                     /// </summary>
                     [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
+                    [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
                         /// <inheritdoc/>
@@ -650,6 +653,7 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                     /// Custom <see cref="IFunctionMetadataProvider"/> implementation that returns function metadata definitions for the current worker."/>
                     /// </summary>
                     [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
+                    [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
                         /// <inheritdoc/>